+++ tinper-bee/bee-datepicker/src/RangePicker.js
@@ -256,10 +256,7 @@ class RangePicker extends Component {
         }
         this.props.onEndInputBlur && this.props.onEndInputBlur(e, endValue, `["${startValue}" , "${endValue}"]`);
     }
-    //阻止组件内部事件冒泡到组件外部容器
-    stopPropagation = (e) => {
-        e.stopPropagation();
-    }
+
     onOk = (value) => {
         this.props.onOk && this.props.onOk(value);
     }
@@ -297,7 +294,7 @@ class RangePicker extends Component {
         />
     );
       return (
-          <div onClick={this.stopPropagation} onMouseOver={this.stopPropagation} 
+          <div
           {...omit(others, [
             'closeIcon',
             'renderIcon',