+++ tinper-bee/bee-loading/src/i18n.js
@@ -1,0 +1,10 @@
+module.exports = {
+    'lang': 'zh-cn',
+    'loading': '加载中...',
+    'en-us': {
+        'loading': 'Loading...',
+    },
+    'zh-tw': {
+        'loading': '加載中...',
+    },
+}