+++ tinper-bee/bee-select/src/RcSelect.js
@@ -623,7 +623,7 @@ class Select extends React.Component {
     if (state.inputValue) {
       hidden = true;
     }
-    if (state.value.length) {
+    if (state.value[0]) {
       hidden = true;
     }
     if (isCombobox(props) && state.value.length === 1 && !state.value[0]) {