+++ tinper-bee/bee-clipboard/CHANGELOG.md
@@ -1,3 +1,12 @@
+## [2.0.5](https://github.com/tinper-bee/bee-clipboard/compare/v2.0.4...v2.0.5) (2020-07-15)
+
+
+### Bug Fixes
+
+* 复制提示文本位置显示错误[[#533](https://github.com/tinper-bee/bee-clipboard/issues/533)] ([ec1586d](https://github.com/tinper-bee/bee-clipboard/commit/ec1586d09edf432aa2c9497c1a5532a088ab76b3))
+
+
+
 ## [2.0.4](https://github.com/tinper-bee/bee-clipboard/compare/v2.0.3...v2.0.4) (2020-04-24)