+++ tinper-bee/bee-modal/CHANGELOG.md
@@ -1,3 +1,26 @@
+## [2.1.18](https://github.com/tinper-bee/bee-modal/compare/v2.1.17...v2.1.18) (2020-07-07)
+
+
+### Features
+
+* needScroll支持传入number类型的值 ([ad6c103](https://github.com/tinper-bee/bee-modal/commit/ad6c1032d64e24194a56ff019c02b60717af25b9))
+
+
+
+## [2.1.17](https://github.com/tinper-bee/bee-modal/compare/v2.1.15...v2.1.17) (2020-07-07)
+
+
+### Bug Fixes
+
+* 去掉 pointer-events 影响 ([f086dd8](https://github.com/tinper-bee/bee-modal/commit/f086dd8b5c20868219fd13403812dddafb45e1a3))
+
+
+### Features
+
+* 新增needScroll属性，解决打开模态框滚动到最顶部的问题 ([7589392](https://github.com/tinper-bee/bee-modal/commit/7589392e2678160ac03aed9b27004b5a8b363a5a))
+
+
+
 ## [2.1.16](https://github.com/tinper-bee/bee-modal/compare/v2.1.15...v2.1.16) (2020-05-28)